+++ stroom-query-api/src/main/java/stroom/datasource/api/v2/FieldTypes.java
@@ -21,14 +21,15 @@ import java.util.Arrays;
 import java.util.List;
 
 public class FieldTypes {
-    public static final String BOOLEAN = "Boolean";
-    public static final String DATE = "Date";
-    public static final String DOC_REF = "DocRef";
     public static final String ID = "Id";
+    public static final String BOOLEAN = "Boolean";
     public static final String INTEGER = "Integer";
     public static final String LONG = "Long";
-    public static final String NUMBER = "Number";
+    public static final String FLOAT = "Float";
+    public static final String DOUBLE = "Double";
+    public static final String DATE = "Date";
     public static final String TEXT = "Text";
+    public static final String DOC_REF = "DocRef";
 
-    public static final List<String> TYPES = new ArrayList<>(Arrays.asList(BOOLEAN, DATE, DOC_REF, ID, INTEGER, LONG, TEXT));
+    public static final List<String> TYPES = new ArrayList<>(Arrays.asList(ID, BOOLEAN, INTEGER, LONG, FLOAT, DOUBLE, DATE, TEXT, DOC_REF));
 }