+++ stroom-query-api/src/test/java/stroom/datasource/api/v2/DataSourceFieldBuilderTest.java
@@ -3,6 +3,7 @@ package stroom.datasource.api.v2;
 import org.junit.jupiter.api.Test;
 import stroom.query.api.v2.ExpressionTerm;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -14,8 +15,10 @@ class DataSourceFieldBuilderTest {
         final String name = "someField";
         final ExpressionTerm.Condition condition = ExpressionTerm.Condition.BETWEEN;
 
+        ArrayList <ExpressionTerm.Condition> conditionArrayList = new ArrayList<>();
+        conditionArrayList.add(condition);
         // When
-        final TextField field = new TextField(name, true, List.of(condition));
+        final TextField field = new TextField(name, true,conditionArrayList);
 
         // Then
         assertThat(field.getConditions()).hasSize(1);