+++ stroom-query-common/src/test/java/stroom/query/common/v2/TestDateExpressionParser.java
@@ -157,6 +157,6 @@ class TestDateExpressionParser {
         } catch (DateTimeException e) {
             dateTimeException = e;
         }
-        Assertions.assertEquals(expectedMessage, dateTimeException.getMessage());
+        assertThat(dateTimeException.getMessage()).isEqualTo(expectedMessage);
     }
 }