+++ stroom-query-testing/src/main/java/stroom/query/testing/DatabaseContainerExtensionSupport.java
@@ -29,7 +29,9 @@ public class DatabaseContainerExtensionSupport implements BeforeAllCallback, Aft
     }
 
     private static Object get(Field member, Object instance) throws IllegalAccessException {
-        if (!member.canAccess(instance)) {
+        //Java 8 does not provide Field.canAccess()
+        //if (!member.canAccess(instance)) {
+        if (true){
             member.setAccessible(true);
         }